knowledge_graph/pipeline/pyproject.toml
[build-system]
requires = ["setuptools==69.0.2"]
build-backend = "setuptools.build_meta"

# We will keep handling dependencies in setup.py
# since it helps us organize and isolate version management
[project]
name = "ailab-kg-pipeline"
version = "0.0.1"
authors = [{ name = "pmbrull" }]
description = "ailab knowledge graph pipeline"
requires-python = ">=3.11"

dependencies = [
    "pydantic~=1.10",
    "neo4j",
    "pyyaml",
    "openmetadata-ingestion",
    "haystack-ai",
    "pgvector-haystack",
    "ollama-haystack",
    "sqlalchemy>=1.4.0,<2",
    # "metadata @ file://${PROJECT_ROOT}/../../projects/OpenMetadata/ingestion",  #  uv pip install "/Users/pmbrull/projects/OpenMetadata/ingestion"
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy~=1.1.1",
    "codespell~=2.2.6",
    "pre-commit",
    "pytest",
    "pyright",
    "jupyterlab",
]

[tool.setuptools.packages.find]
where = ["."] # list of folders that contain the packages (["."] by default)

[tool.ruff]
line-length = 120
extend-select = ['Q', 'RUF100', 'C90', 'UP', 'I', 'D', 'T']
extend-ignore = ['D105', 'D107', 'D205', 'D415']
flake8-quotes = { inline-quotes = 'double', multiline-quotes = 'double' }
mccabe = { max-complexity = 14 }
isort = { known-first-party = ['server', 'tests'] }
target-version = "py311"

[tool.ruff.extend-per-file-ignores]
"tests/**/*.py" = ['T', 'E721', 'F811']

[tool.ruff.format]
quote-style = "double"

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.per-file-ignores]
'tests/*' = ['D']
"__init__.py" = ['D']

[tool.codespell]
skip = ".git,env*,pydantic/v1/*,*.java"

[tool.pyright]
venvPath = "."
venv = ".venv"
